create class
The program creates a callable task that generates a collection of 10 random integers and falls asleep randomly for 1-10 seconds, outputs the sum of these numbers as a string. Runs 10 tasks in parallel in a pool of 3 threads. Prints the progress of the program to the screen, indicating the name of the thread that is doing the work.