create new thread
Program creates a callable task that takes a HELLO WORLD message, the current time, and writes it to a file. Writing to a file is done sequentially via a synchronized method. Run 10 tasks in parallel in a pool of 3 threads. Prints the progress of the program to the screen, indicating the name of the thread that is doing the work.